fix: add build tags so code compiles on macOS/Windows
This moves the Linux-specific syscalls (sched_setaffinity, CPUSet) into
platform-specific files with build tags. Now developers on macOS can
run the linter and tests locally without everything exploding.

The actual affinity functionality still only works on Linux (as it should -
this is a Proxmox tool after all), but at least `go build` and `go test`
work everywhere now.
