+++ pkg/scheduler/affinity_test.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
-	"golang.org/x/sys/unix"
 
 	"github.com/egandro/proxmox-cpu-affinity/pkg/config"
 	"github.com/egandro/proxmox-cpu-affinity/pkg/cpuinfo"
@@ -36,7 +35,7 @@ type MockSystemAffinityOps struct {
 	mock.Mock
 }
 
-func (m *MockSystemAffinityOps) SchedSetaffinity(pid int, mask *unix.CPUSet) error {
+func (m *MockSystemAffinityOps) SchedSetaffinity(pid int, mask *CPUSet) error {
 	args := m.Called(pid, mask)
 	return args.Error(0)
 }
@@ -109,7 +108,7 @@ func TestApplyAffinity(t *testing.T) {
 			setupMockSys: func(m *MockSystemAffinityOps) {
 				m.On("GetProcessThreads", 12345).Return([]int{12345}, nil)
 				m.On("GetChildProcesses", 12345).Return([]int{}, nil)
-				m.On("SchedSetaffinity", 12345, mock.MatchedBy(func(mask *unix.CPUSet) bool {
+				m.On("SchedSetaffinity", 12345, mock.MatchedBy(func(mask *CPUSet) bool {
 					return mask.IsSet(1) && mask.IsSet(0) && !mask.IsSet(2)
 				})).Return(nil)
 			},
